Work on typings for image and artwork
