+++ src/components/publishing/image.tsx
@@ -1,46 +1,22 @@
 import React from "react"
 import styled from "styled-components"
-import Fonts from "./fonts"
-
-const Figcaption = styled.div`
-  & > p {
-    ${Fonts.unica("s14", "medium")}
-    margin: 0;
-    color: #999;
-  }
-`
-
-const CaptionLink = styled.div`
-  ${Fonts.unica("s14", "medium")}
-  margin: 0;
-  margin-left: 10px;
-  border-bottom: 1px solid black;
-  cursor: pointer;
-  display: inline-block;
-  min-width: 7.1em;
-  white-space: nowrap;
-  align-self: flex-start;
-`
-
-const CaptionContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  margin: 10px 0 10px 0;
-`
+import ImageCaption from "./image_caption"
 
 const BlockImage = styled.img`
   display: block;
 `
 
-function Image(props) {
-  const { image } = props
+interface ImageProps extends React.HTMLProps<HTMLDivElement> {
+  image?: any
+  layout?: string
+}
+
+const Image: React.SFC<ImageProps> = props => {
+  const { image, layout } = props
   return (
     <div className="article-image">
       <BlockImage src={image.url} width="100%" />
-      <CaptionContainer>
-        <Figcaption dangerouslySetInnerHTML={{ __html: image.caption }} />
-        <CaptionLink>View Fullscreen</CaptionLink>
-      </CaptionContainer>
+      <ImageCaption image={image} layout={layout} />
     </div>
   )
 }
